integrated bluetooth and altered partitions
Bluetooth connectivity with phone and laptop were shown to work. However, to reduce space the BLEkeyboard library was reduced to just concern media keys, not the entire keyboard, this may be reverted later. Having both NimBLE, Wifi and GFX present on one device used a lot of flash space. Therefore the app0 partition size was increased from 1.3MB to 2MB. Given that no more libraries are required. This should allow for enough space for the rest of the code